Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.post1
pymongo==4.6.0
python-decouple==3.8
